% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Load the data files in a directory.}
\usage{
compile(data_dir, regex_key, header_file)
}
\arguments{
\item{data_dir}{The directory of the data files, passed to `file.path`.}

\item{regex_key}{Pattern passed to `list.files`. Matches are read.}

\item{header_file}{A file containing the data column names. Assumed
to be in data_dir, unless it exists.}
}
\value{
dplyr::data_frame
}
\description{
Load the data files in a directory.
}

